Allow re-using extracted content
Usage:
export TMPDIR=/some/dir/$device
export SKIP_CLEANUP=true

+++ build/tools/extract_utils.sh
@@ -28,7 +28,8 @@ COMMON=-1
 ARCHES=
 FULLY_DEODEXED=-1
 
-TMPDIR=$(mktemp -d)
+SKIP_CLEANUP=${SKIP_CLEANUP:-0}
+TMPDIR=${TMPDIR:-$(mktemp -d)}
 HOST="$(uname | tr '[:upper:]' '[:lower:]')"
 
 #
@@ -37,7 +38,11 @@ HOST="$(uname | tr '[:upper:]' '[:lower:]')"
 # kill our tmpfiles with fire on exit
 #
 function cleanup() {
-    rm -rf "${TMPDIR:?}"
+    if [ "$SKIP_CLEANUP" == "true" ] || [ "$SKIP_CLEANUP" == "1" ]; then
+        echo "Skipping cleanup of $TMPDIR"
+    else
+        rm -rf "${TMPDIR:?}"
+    fi
 }
 
 trap cleanup 0
